requirements.txt
Django==5.0.2
python-dotenv==1.0.1
Pillow==10.4.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
requests==2.31.0
Djasgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.7
cssselect2==0.7.0
defusedxml==0.7.1
fonttools==4.46.0
html5lib==1.1
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
lxml==4.9.3
matplotlib==3.8.2
numpy==1.26.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.8.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
reportlab==4.0.7
selenium==4.15.2
selenium-wire==5.1.0
soupsieve==2.5
sqlparse==0.4.4
tinycss2==1.2.1
tqdm==4.66.1
urllib3==2.1.0
weasyprint==60.2
webencodings==0.5.1
